requirements.txt
# MedGemma 1.5 胸部 X 光报告生成 - 依赖
# Python 3.10-3.12

# 核心
torch>=2.6.0
torchvision>=0.21.0
transformers>=5.0.0
accelerate
pillow
pandas

# 评估
radgraph

# 量化（W4A4 / W4A8）
bitsandbytes>=0.46.1

# AutoRound 4bit
auto-gptq
optimum

# 蒸馏
peft
trl
